Remove contents of symlink to dir with rm -rf (#688)
* rm -rf on a symlink to a dir deletes contents

* Fix comment typo

* Clarify comments

* Pass symlink directly to rmdirSync

* Skip rm -rf symlink test on windows

+++ test/rm.js
@@ -150,7 +150,7 @@ test('recursive dir removal', t => {
   t.is(result.code, 0);
   const contents = fs.readdirSync(t.context.tmp);
   t.is(contents.length, 1);
-  t.is(contents[0], '.hidden'); // shouldn't remove hiddden if no .* given
+  t.is(contents[0], '.hidden'); // shouldn't remove hidden if no .* given
 });
 
 test('recursive dir removal #2', t => {
@@ -267,16 +267,17 @@ test('remove symbolic link to a dir', t => {
   t.truthy(fs.existsSync(`${t.context.tmp}/rm/a_dir`));
 });
 
-// TODO(nfischer): fix this behavior in rm
-test.skip('rm -rf on a symbolic link to a dir deletes its contents', t => {
-  const result = shell.rm('-rf', `${t.context.tmp}/rm/link_to_a_dir/`);
-  t.falsy(shell.error());
-  t.is(result.code, 0);
+test('rm -rf on a symbolic link to a dir deletes its contents', t => {
+  if (process.platform !== 'win32') {
+    const result = shell.rm('-rf', `${t.context.tmp}/rm/link_to_a_dir/`);
+    t.falsy(shell.error());
+    t.is(result.code, 0);
 
-  // Both the link and original dir should remain, but contents are deleted
-  t.truthy(fs.existsSync(`${t.context.tmp}/rm/link_to_a_dir`));
-  t.truthy(fs.existsSync(`${t.context.tmp}/rm/a_dir`));
-  t.falsy(fs.existsSync(`${t.context.tmp}/rm/a_dir/a_file`));
+    // Both the link and original dir should remain, but contents are deleted
+    t.truthy(fs.existsSync(`${t.context.tmp}/rm/link_to_a_dir`));
+    t.truthy(fs.existsSync(`${t.context.tmp}/rm/a_dir`));
+    t.falsy(fs.existsSync(`${t.context.tmp}/rm/a_dir/a_file`));
+  }
 });
 
 test('remove broken symbolic link', t => {
